Extract collapse Compute class
