+++ tests/line/__init__.py
@@ -1,3 +1,2 @@
-from .combinations import CombinationsShould
 from .taskline import TaskLineShould
 from .fieldline import FieldLineShould
